Fix race on reading stats (from race detector - thanks WillN).
Pulled the counters into an AppStats structure, turned into
a lexical in the main goro to ensure encapsulation.

Added a GetStats() method on the app which returns the data.

Internally implemented as a requestor "chan chan AppStats". Where
you send a "chan AppStats" in, and the main loop writes back the
data then closes your request channel.

One subtlety is that you can't call GetStats() from the dynamic
scope of the main goro. I had a bug where the stats usage in
"finished()" was calling GetStats() which deadlocked. Fixed by
passing the AppStats down directly to finished().