Add gopctl cmd line and a client lib.
A gop.Client for talking to the gop HTTP API under /gop and a command line tool
for grabbing status, stacks, setting config etc. See gopctl -h.

Something thats been brewing on the side for while, has been really helpful
tracking down probs in running silkis and admirals, so pushing.
Still not sure on the right cmd line API for specifing the app to talk to.